+++ packages/experimental/Avatar/src/JSAvatar.styling.ts
@@ -11,7 +11,17 @@ const nameMap: { [key: string]: string } = {
   end: 'flex-end',
 };
 
-export const avatarStates: (keyof JSAvatarTokens)[] = [...AvatarColors, ...AvatarSizesForTokens, 'circular', 'square', 'inactive'];
+export const avatarStates: (keyof JSAvatarTokens)[] = [
+  ...AvatarColors,
+  ...AvatarSizesForTokens,
+  'circular',
+  'square',
+  'inactive',
+  'ringColor',
+  'ringBackgroundColor',
+  'iconColor',
+  'size',
+];
 
 export const stylingSettings: UseStylingOptions<JSAvatarProps, AvatarSlotProps, JSAvatarTokens> = {
   tokens: [defaultJSAvatarTokens, JSAvatarName],
@@ -19,12 +29,12 @@ export const stylingSettings: UseStylingOptions<JSAvatarProps, AvatarSlotProps, 
   slotProps: {
     root: buildProps(
       (tokens: JSAvatarTokens) => {
-        const { horizontalIconAlignment, verticalIconAlignment, width, height, avatarOpacity } = tokens;
+        const { horizontalIconAlignment, verticalIconAlignment, size, avatarOpacity } = tokens;
         return {
           style: {
             flexDirection: 'row',
-            width: width,
-            height: height,
+            width: size,
+            height: size,
             justifyContent: nameMap[horizontalIconAlignment || 'end'] as ViewStyle['justifyContent'],
             alignItems: nameMap[verticalIconAlignment || 'end'] as ViewStyle['alignItems'],
             horizontalIconAlignment,
@@ -33,7 +43,7 @@ export const stylingSettings: UseStylingOptions<JSAvatarProps, AvatarSlotProps, 
           },
         };
       },
-      ['horizontalIconAlignment', 'verticalIconAlignment', 'avatarOpacity', 'height', 'width'],
+      ['horizontalIconAlignment', 'verticalIconAlignment', 'avatarOpacity', 'size'],
     ),
     initials: buildProps(
       (tokens: JSAvatarTokens, theme: Theme) => {
@@ -41,6 +51,7 @@ export const stylingSettings: UseStylingOptions<JSAvatarProps, AvatarSlotProps, 
           style: {
             ...fontStyles.from(tokens, theme),
             color: tokens.color,
+            textAlign: 'center',
           },
         };
       },
@@ -48,64 +59,71 @@ export const stylingSettings: UseStylingOptions<JSAvatarProps, AvatarSlotProps, 
     ),
     initialsBackground: buildProps(
       (tokens: JSAvatarTokens, theme: Theme) => {
-        const { backgroundColor } = tokens;
+        const { backgroundColor, size } = tokens;
 
         return {
           style: {
             ...borderStyles.from(tokens, theme),
-            width: tokens.width,
-            height: tokens.height,
+            width: size,
+            height: size,
             flexGrow: 1,
             alignSelf: 'stretch',
             justifyContent: 'center',
             alignItems: 'center',
             backgroundColor: backgroundColor,
+            borderWidth: tokens.borderWidth,
+            borderColor: tokens.borderColor,
           },
         };
       },
-      ['backgroundColor', 'width', 'height', ...borderStyles.keys],
+      ['backgroundColor', 'size', 'borderColor', 'borderWidth', ...borderStyles.keys],
     ),
     image: buildProps(
       (tokens: JSAvatarTokens) => {
+        const { borderRadius, size, borderWidth, borderColor } = tokens;
         return {
           style: {
-            borderRadius: tokens.borderRadius,
-            width: tokens.width,
-            height: tokens.height,
+            borderRadius: borderRadius,
+            width: size,
+            height: size,
+            borderWidth: borderWidth,
+            borderColor: borderColor,
           },
         };
       },
-      ['borderRadius', 'width', 'height'],
+      ['borderRadius', 'size', 'borderColor', 'borderWidth'],
     ),
     icon: buildProps(
       (tokens: JSAvatarTokens) => {
         return {
           style: {
             position: 'absolute',
+            color: tokens.iconColor,
             width: tokens.iconSize,
             height: tokens.iconSize,
           },
         };
       },
-      ['iconSize'],
+      ['iconSize', 'iconColor'],
     ),
     ring: buildProps(
       (tokens: JSAvatarTokens, theme: Theme) => {
-        const ringConfig = getRingConfig(tokens.width);
+        const ringConfig = getRingConfig(tokens.size);
         return {
           style: {
             borderStyle: 'solid',
-            borderWidth: ringConfig.stroke,
             width: ringConfig.size,
             height: ringConfig.size,
             position: 'absolute',
             top: -ringConfig.stroke * 2,
             left: -ringConfig.stroke * 2,
             ...borderStyles.from(tokens, theme),
+            borderWidth: ringConfig.stroke,
+            borderColor: tokens.ringColor,
           },
         };
       },
-      ['width', 'height', ...borderStyles.keys],
+      ['size', 'ringColor', ...borderStyles.keys],
     ),
     badge: buildProps(
       (tokens: JSAvatarTokens) => {