+++ packages/experimental/Avatar/src/JSAvatarTokens.ts
@@ -13,8 +13,12 @@ export const defaultJSAvatarTokens: TokenSettings<JSAvatarTokens, Theme> = (t: T
     fontFamily: t.typography.families.primary,
     fontWeight: globalTokens.font.weight.semibold,
     fontSize: globalTokens.font.size[200],
-    width: 24,
-    height: 24,
+    size: 24,
+    iconSize: 16,
+    iconColor: globalTokens.color.white,
+    ringColor: t.colors.transparentStroke,
+    borderColor: globalTokens.color.white,
+    borderWidth: t.host.appearance === 'highContrast' ? 1 : 0,
     circular: {
       borderRadius: globalTokens.corner.radius.circle,
     },
@@ -25,8 +29,7 @@ export const defaultJSAvatarTokens: TokenSettings<JSAvatarTokens, Theme> = (t: T
       avatarOpacity: 0.8,
     },
     size20: {
-      width: 20,
-      height: 20,
+      size: 20,
       badgeSize: 'smallest',
       iconSize: 16,
       fontSize: globalTokens.font.size[100],
@@ -35,8 +38,7 @@ export const defaultJSAvatarTokens: TokenSettings<JSAvatarTokens, Theme> = (t: T
       },
     },
     size24: {
-      width: 24,
-      height: 24,
+      size: 24,
       badgeSize: 'smallest',
       iconSize: 16,
       fontSize: globalTokens.font.size[100],
@@ -45,41 +47,35 @@ export const defaultJSAvatarTokens: TokenSettings<JSAvatarTokens, Theme> = (t: T
       },
     },
     size28: {
-      width: 28,
-      height: 28,
+      size: 28,
       badgeSize: 'smaller',
       iconSize: 20,
       fontWeight: globalTokens.font.weight.semibold,
       fontSize: globalTokens.font.size[100],
     },
     size32: {
-      width: 32,
-      height: 32,
+      size: 32,
       badgeSize: 'smaller',
       iconSize: 20,
       fontSize: globalTokens.font.size[100],
     },
     size36: {
-      width: 36,
-      height: 36,
+      size: 36,
       badgeSize: 'smaller',
       iconSize: 20,
     },
     size40: {
-      width: 40,
-      height: 40,
+      size: 40,
       badgeSize: 'small',
       iconSize: 20,
     },
     size48: {
-      width: 48,
-      height: 48,
+      size: 48,
       badgeSize: 'small',
       iconSize: 24,
     },
     size56: {
-      width: 56,
-      height: 56,
+      size: 56,
       badgeSize: 'medium',
       iconSize: 28,
       fontSize: globalTokens.font.size[400],
@@ -88,8 +84,7 @@ export const defaultJSAvatarTokens: TokenSettings<JSAvatarTokens, Theme> = (t: T
       },
     },
     size64: {
-      width: 64,
-      height: 64,
+      size: 64,
       badgeSize: 'large',
       iconSize: 32,
       fontSize: globalTokens.font.size[500],
@@ -98,8 +93,7 @@ export const defaultJSAvatarTokens: TokenSettings<JSAvatarTokens, Theme> = (t: T
       },
     },
     size72: {
-      width: 72,
-      height: 72,
+      size: 72,
       badgeSize: 'large',
       iconSize: 32,
       fontSize: globalTokens.font.size[500],
@@ -108,8 +102,7 @@ export const defaultJSAvatarTokens: TokenSettings<JSAvatarTokens, Theme> = (t: T
       },
     },
     size96: {
-      width: 96,
-      height: 96,
+      size: 96,
       badgeSize: 'largest',
       iconSize: 48,
       fontWeight: globalTokens.font.weight.regular,
@@ -119,8 +112,7 @@ export const defaultJSAvatarTokens: TokenSettings<JSAvatarTokens, Theme> = (t: T
       },
     },
     size120: {
-      width: 120,
-      height: 120,
+      size: 120,
       badgeSize: 'largest',
       iconSize: 48,
       fontWeight: globalTokens.font.weight.regular,
@@ -132,161 +124,77 @@ export const defaultJSAvatarTokens: TokenSettings<JSAvatarTokens, Theme> = (t: T
     neutral: {
       backgroundColor: t.colors.neutralBackground6,
       color: t.colors.neutralForeground3,
-      borderColor: t.colors.defaultBorder,
+      iconColor: t.colors.neutralForeground3,
+      ringColor: t.colors.transparentStroke,
     },
     brand: {
       backgroundColor: t.colors.brandBackgroundStatic,
       color: t.colors.neutralForegroundOnBrand,
-      borderColor: t.colors.brandedBorder,
+      iconColor: t.colors.neutralForegroundOnBrand,
+      ringColor: t.colors.transparentStroke,
     },
-    darkRed: {
-      backgroundColor: globalTokens.color.darkRed.tint40,
-      color: globalTokens.color.darkRed.shade30,
-      borderColor: globalTokens.color.darkRed.primary,
-    },
-    cranberry: {
-      backgroundColor: globalTokens.color.cranberry.tint40,
-      color: globalTokens.color.cranberry.shade30,
-      borderColor: globalTokens.color.cranberry.primary,
-    },
-    red: {
-      backgroundColor: globalTokens.color.red.tint40,
-      color: globalTokens.color.red.shade30,
-      borderColor: globalTokens.color.red.primary,
-    },
-    pumpkin: {
-      backgroundColor: globalTokens.color.pumpkin.tint40,
-      color: globalTokens.color.pumpkin.shade30,
-      borderColor: globalTokens.color.pumpkin.primary,
-    },
-    peach: {
-      backgroundColor: globalTokens.color.peach.tint40,
-      color: globalTokens.color.peach.shade30,
-      borderColor: globalTokens.color.peach.primary,
-    },
-    marigold: {
-      backgroundColor: globalTokens.color.marigold.tint40,
-      color: globalTokens.color.marigold.shade30,
-      borderColor: globalTokens.color.marigold.primary,
-    },
-    gold: {
-      backgroundColor: globalTokens.color.gold.tint40,
-      color: globalTokens.color.gold.shade30,
-      borderColor: globalTokens.color.gold.primary,
-    },
-    brass: {
-      backgroundColor: globalTokens.color.brass.tint40,
-      color: globalTokens.color.brass.shade30,
-      borderColor: globalTokens.color.brass.primary,
-    },
-    brown: {
-      backgroundColor: globalTokens.color.brown.tint40,
-      color: globalTokens.color.brown.shade30,
-      borderColor: globalTokens.color.brown.primary,
-    },
-    forest: {
-      backgroundColor: globalTokens.color.forest.tint40,
-      color: globalTokens.color.forest.shade30,
-      borderColor: globalTokens.color.forest.primary,
-    },
-    seafoam: {
-      backgroundColor: globalTokens.color.seafoam.tint40,
-      color: globalTokens.color.seafoam.shade30,
-      borderColor: globalTokens.color.seafoam.primary,
-    },
-    darkGreen: {
-      backgroundColor: globalTokens.color.darkGreen.tint40,
-      color: globalTokens.color.darkGreen.shade30,
-      borderColor: globalTokens.color.darkGreen.primary,
-    },
-    lightTeal: {
-      backgroundColor: globalTokens.color.lightTeal.tint40,
-      color: globalTokens.color.lightTeal.shade30,
-      borderColor: globalTokens.color.lightTeal.primary,
-    },
-    teal: {
-      backgroundColor: globalTokens.color.teal.tint40,
-      color: globalTokens.color.teal.shade30,
-      borderColor: globalTokens.color.teal.primary,
-    },
-    steel: {
-      backgroundColor: globalTokens.color.steel.tint40,
-      color: globalTokens.color.steel.shade30,
-      borderColor: globalTokens.color.steel.primary,
-    },
-    blue: {
-      backgroundColor: globalTokens.color.blue.tint40,
-      color: globalTokens.color.blue.shade30,
-      borderColor: globalTokens.color.blue.primary,
-    },
-    royalBlue: {
-      backgroundColor: globalTokens.color.royalBlue.tint40,
-      color: globalTokens.color.royalBlue.shade30,
-      borderColor: globalTokens.color.royalBlue.primary,
-    },
-    cornflower: {
-      backgroundColor: globalTokens.color.cornflower.tint40,
-      color: globalTokens.color.cornflower.shade30,
-      borderColor: globalTokens.color.cornflower.primary,
-    },
-    navy: {
-      backgroundColor: globalTokens.color.navy.tint40,
-      color: globalTokens.color.navy.shade30,
-      borderColor: globalTokens.color.navy.primary,
-    },
-    lavender: {
-      backgroundColor: globalTokens.color.lavender.tint40,
-      color: globalTokens.color.lavender.shade30,
-      borderColor: globalTokens.color.lavender.primary,
-    },
-    purple: {
-      backgroundColor: globalTokens.color.purple.tint40,
-      color: globalTokens.color.purple.shade30,
-      borderColor: globalTokens.color.purple.primary,
-    },
-    grape: {
-      backgroundColor: globalTokens.color.grape.tint40,
-      color: globalTokens.color.grape.shade30,
-      borderColor: globalTokens.color.grape.primary,
-    },
-    lilac: {
-      backgroundColor: globalTokens.color.lilac.tint40,
-      color: globalTokens.color.lilac.shade30,
-      borderColor: globalTokens.color.lilac.primary,
-    },
-    pink: {
-      backgroundColor: globalTokens.color.pink.tint40,
-      color: globalTokens.color.pink.shade30,
-      borderColor: globalTokens.color.pink.primary,
-    },
-    magenta: {
-      backgroundColor: globalTokens.color.magenta.tint40,
-      color: globalTokens.color.magenta.shade30,
-      borderColor: globalTokens.color.magenta.primary,
-    },
-    plum: {
-      backgroundColor: globalTokens.color.plum.tint40,
-      color: globalTokens.color.plum.shade30,
-      borderColor: globalTokens.color.plum.primary,
-    },
-    beige: {
-      backgroundColor: globalTokens.color.beige.tint40,
-      color: globalTokens.color.beige.shade30,
-      borderColor: globalTokens.color.beige.primary,
-    },
-    mink: {
-      backgroundColor: globalTokens.color.mink.tint40,
-      color: globalTokens.color.mink.shade30,
-      borderColor: globalTokens.color.mink.primary,
-    },
-    platinum: {
-      backgroundColor: globalTokens.color.platinum.tint40,
-      color: globalTokens.color.platinum.shade30,
-      borderColor: globalTokens.color.platinum.primary,
-    },
-    anchor: {
-      backgroundColor: globalTokens.color.anchor.tint40,
-      color: globalTokens.color.anchor.shade30,
-      borderColor: globalTokens.color.anchor.primary,
-    },
+    darkRed: getColorProps('darkRed', t),
+    cranberry: getColorProps('cranberry', t),
+    red: getColorProps('red', t),
+    pumpkin: getColorProps('pumpkin', t),
+    peach: getColorProps('peach', t),
+    marigold: getColorProps('marigold', t),
+    gold: getColorProps('gold', t),
+    brass: getColorProps('brass', t),
+    brown: getColorProps('brown', t),
+    forest: getColorProps('forest', t),
+    seafoam: getColorProps('seafoam', t),
+    darkGreen: getColorProps('darkGreen', t),
+    lightTeal: getColorProps('lightTeal', t),
+    teal: getColorProps('teal', t),
+    steel: getColorProps('steel', t),
+    blue: getColorProps('blue', t),
+    royalBlue: getColorProps('royalBlue', t),
+    cornflower: getColorProps('cornflower', t),
+    navy: getColorProps('navy', t),
+    lavender: getColorProps('lavender', t),
+    purple: getColorProps('purple', t),
+    grape: getColorProps('grape', t),
+    lilac: getColorProps('lilac', t),
+    pink: getColorProps('pink', t),
+    magenta: getColorProps('magenta', t),
+    plum: getColorProps('plum', t),
+    beige: getColorProps('beige', t),
+    mink: getColorProps('mink', t),
+    platinum: getColorProps('platinum', t),
+    anchor: getColorProps('anchor', t),
   } as JSAvatarTokens);
+
+/**
+ * A function which returns object of props depending on color and theme.
+ * @param color
+ * @param theme
+ * @returns object of props - backgroundColor, color and ringColor
+ */
+function getColorProps(color: string, theme: Theme) {
+  const themeAppearance = theme.host.appearance;
+  switch (themeAppearance) {
+    case 'light':
+    default:
+      return {
+        backgroundColor: globalTokens.color[color].tint40,
+        color: globalTokens.color[color].shade30,
+        iconColor: globalTokens.color[color].shade30,
+        ringColor: globalTokens.color[color].primary,
+      };
+    case 'dark':
+      return {
+        backgroundColor: globalTokens.color[color].shade30,
+        color: globalTokens.color[color].tint40,
+        iconColor: globalTokens.color[color].tint40,
+        ringColor: globalTokens.color[color].tint30,
+      };
+    case 'highContrast':
+      return {
+        backgroundColor: theme.colors.neutralBackground6,
+        color: theme.colors.neutralForeground3,
+        iconColor: theme.colors.neutralForeground3,
+        ringColor: theme.colors.transparentStroke,
+      };
+  }
+}