+++ packages/theming/theming-utils/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui-react-native/theming-utils
 
-This log was last generated on Thu, 28 Apr 2022 19:09:51 GMT and should not be manually modified.
+This log was last generated on Sat, 21 May 2022 01:47:57 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.11.0
+
+Sat, 21 May 2022 01:47:57 GMT
+
+### Minor changes
+
+- Add shadows to the theme (ruaraki@microsoft.com)
+- Bump @fluentui-react-native/theme-types to v0.15.0
 
 ## 0.10.2